refactor(nala): move `write_history/log` after `--download-only` check
By moving these two closer to the start of `dpkg` it ensures we don't get `--download-only` transactions into the log or history file.

Since we are already checking for that switch in the main file we can remove it from `write_history` function.

There is an argument where these maybe should be after `start.dpkg` but if we do that and anything happens during the processing of packages we won't get a log or history.