fix(downloader): redownloading packages that already exist
To fix this issue we process each downloaded file right after it's downloaded. SO if the `.deb` is good, it immediately is moved to the Archive Directory. This behavior more closely mimics `apt`.

Previously we were processing all the downloads at the end. Additionally doing it this way if a processed file fails we can try to download again.